pyproject.toml
[project]
name = "cerebro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "onnxruntime>=1.23.2",
    "pillow>=12.1.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "boto3",
    "requests",
    "ultralytics",
    "onnx>=1.20.1",
    "onnxslim>=0.1.83",
    "pytest-benchmark>=4.0.0",
    "psutil>=5.9.0",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["detection*"]
namespaces = false
